fix: perform deep equality check on json config
Compare the request and resource JSON configuration structures using a
deep equality check, rather than a byte equality check, as
(de)serialization will change the structure (e.g., by removing
whitespace).

Also be sure to check for expiry by the current UTC timestamp, as that
is how the expiry is stored in the state file.