Add github app rotator
Github uses rate limiting for requests, when the limit is reached a
request returns a RateLimitError. The rate limit is per app client.
This commit adds the ability to use multiple github apps, and rotate
between them for calls to github, to lessen the chances of getting
rate limit errors.